requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=openblas
blosc=1.21.0=h4ff587b_1
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.10.11=h06a4308_0
cairo=1.16.0=h19f5f5c_2
certifi=2022.9.24=py310h06a4308_0
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=2.1.1=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
contourpy=1.0.5=py310hdb19cb5_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.12.0=py310h5eee18b_0
dask-core=2022.7.0=py310h06a4308_0
dbus=1.13.18=hb2f20db_0
eigen=3.3.7=hd09550d_1
expat=2.4.9=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
fftw=3.3.9=h27cfd23_1
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fsspec=2022.11.0=py310h06a4308_0
geos=3.8.0=he6710b0_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
harfbuzz=4.3.0=hd55b92a_0
hdf5=1.10.6=h3ffc7dd_1
icu=58.2=he6710b0_3
idna=3.4=pypi_0
imagecodecs=2021.8.26=py310hecf7e94_1
imageio=2.19.3=py310h06a4308_0
imgaug=0.4.0=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.1=py310h06a4308_0
jpeg=9e=h7f8727e_0
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.2=py310h295c915_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libcurl=7.86.0=h91b91d3_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=hbcb73fb_5
libnghttp2=1.46.0=hce63b2e_0
libopenblas=0.3.21=h043d6bf_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
libzopfli=1.0.3=he6710b0_0
locket=1.0.0=py310h06a4308_0
lz4-c=1.9.3=h295c915_1
matplotlib-base=3.6.2=py310h945d387_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py310h06a4308_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.4=py310h757a811_0
numpy=1.23.4=py310hac523dd_0
numpy-base=1.23.4=py310h375b286_0
opencv=4.6.0=py310hefb4dc4_2
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.2=py310h1128e8f_0
partd=1.2.0=pyhd3eb1b0_1
pcre=8.45=h295c915_0
pillow=9.3.0=pypi_0
pip=22.3.1=py310h06a4308_0
pixman=0.40.0=h7f8727e_1
pyparsing=3.0.9=py310h06a4308_0
python=3.10.8=h7a1cb2a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.1=py310h06a4308_0
pywavelets=1.4.1=py310h5eee18b_0
pyyaml=6.0=py310h5eee18b_1
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.1=pypi_0
scikit-image=0.19.3=py310h6a678d5_1
scikit-learn=1.1.3=py310h6a678d5_0
scipy=1.9.3=py310heeff2f4_0
setuptools=65.5.0=py310h06a4308_0
shapely=1.8.4=py310h81ba7c5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.40.0=h5082296_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.12=h1ccaba5_0
toolz=0.12.0=py310h06a4308_0
torch=1.13.0+rocm5.2=pypi_0
torchaudio=0.13.0+rocm5.2=pypi_0
torchvision=0.14.0+rocm5.2=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.13=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zfp=0.5.5=h295c915_6
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
